clean ups and Updates
