+++ backend/src/mms.Application/Common/Mapper/AutoMapperInitializer.cs
@@ -8,7 +8,6 @@ using UserPrivacyEntity = mms.Domain.Entities.UserPrivacy;
 using SupportEntity = mms.Domain.Entities.Support;
 using ProgrammeEntity = mms.Domain.Entities.Programme;
 using mms.Application.Support.Command;
-using mms.Application.Program.Query;
 using AppUserEntity = mms.Domain.Entities.AppUser;
 using FAQEntity = mms.Domain.Entities.FAQ;
 using mms.Application.FAQ.Command;
