+++ backend/src/mms.Application/Programs/Query/GetArchiveProgramsCommand.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using AspNetCoreHero.Results;
+using MediatR;
+using mms.Application.UserNotification.Query;
+
+namespace mms.Application.Programme.Query
+{
+    public class GetArchiveProgramsCommand : IRequest<IResult<List<GetArchiveProgramsResponse>>>
+    {
+    }
+}
+
