+++ backend/src/mms.Application/Programs/Query/GetArchiveProgramsCommandHandler.cs
@@ -6,9 +6,9 @@ using mms.Application.UserNotification.Query;
 using mms.Infrastructure.Context;
 using mms.Infrastructure.Interface;
 
-namespace mms.Application.Program.Query
+namespace mms.Application.Programme.Query
 {
-	public class GetArchiveProgramsCommandHandler : IRequestHandler<GetArchiveProgramsCommand,
+    public class GetArchiveProgramsCommandHandler : IRequestHandler<GetArchiveProgramsCommand,
             IResult<List<GetArchiveProgramsResponse>>>
     {
         private readonly ICurrentUserService _currentUserService;
