+++ backend/src/mms.Application/mms.Application.csproj
@@ -29,7 +29,6 @@
     <None Remove="Support\" />
     <None Remove="Support\Command\" />
     <None Remove="Program\" />
-    <None Remove="Program\Query\" />
     <None Remove="FAQ\" />
     <None Remove="FAQ\Command\" />
     <None Remove="FAQ\Query\" />
@@ -39,8 +38,6 @@
     <Folder Include="Account\ChangePassword\" />
     <Folder Include="Support\" />
     <Folder Include="Support\Command\" />
-    <Folder Include="Program\" />
-    <Folder Include="Program\Query\" />
     <Folder Include="FAQ\" />
     <Folder Include="FAQ\Command\" />
     <Folder Include="FAQ\Query\" />
